+++ src/pages/index.tsx
@@ -1,24 +1,23 @@
-import { Inter } from 'next/font/google'
-import Head from 'next/head'
-import TaskList from '@/pages/components/TaskList'
-import styles from '@/styles/Home.module.css'
+import { Inter as interFont } from 'next/font/google';
+import Head from 'next/head';
+import TaskList from '@/components/TaskList';
 
-const inter = Inter({ subsets: ['latin'] })
+const inter = interFont({ subsets: ['latin'] });
 
-const Home = () =>{
-    return (
-        <>
-        <Head>
-            <meta name="description" content="Generated by create next app" />
-            <meta name="viewport" content="width=device-width, initial-scale=1" />
-            <link rel="icon" href="/favicon.ico" />
-        </Head>
-        <main className={`${styles.main} ${inter.className}`}>
+const Home = () => {
+  return (
+    <>
+      <Head>
+        <meta name="description" content="Generated by create next app" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <main className={`${inter.className}`}>
         <div>
-        <TaskList />
+          <TaskList />
         </div>
-        </main>       
-        </>        
-    )
-}
+      </main>
+    </>
+  );
+};
 export default Home;
